+++ CHANGELOG.md
@@ -507,7 +507,7 @@ with pytest.raises(TypeError):
 ```
 
 
-## version 1.1.8
+## version 1.2.0
 
 improvements
 
@@ -568,4 +568,37 @@ def test_graph_partial_resolve_with_dep():
     data = dg.resolve(Data, name="data")
     assert data.name == "data"
     assert data.db is DATABASE
+```
+
+
+## version 1.2.1
+
+a quick bug fix
+
+```py
+from ididi import DependencyGraph
+
+
+class Time:
+    def __init__(self, name: str):
+        self.name = name
+
+
+def test_registered_singleton():
+    """
+    previously
+    dg.should_be_scoped(dep_type) would analyze dependencies of the
+    dep_type, and if it contains resolvable dependencies exception would be raised
+
+    since registered_singleton is managed by user,
+    dg.should_be_scoped(registered_singleton) should always be False.
+    """
+    timer = Time("1")
+
+    dg = DependencyGraph()
+
+    dg.register_singleton(timer)
+
+    assert dg.should_be_scoped(Time) is False
+
 ```